Code cleanup and amending varying mp3-128 links
As well as realisation that Bandcamp serves the mp3 from the mp3-128 link with any tags the mp3 was uploaded with, so the get_tags and set_tags functions would have been completely redundant.

I might re-add them in the future if I end up making checks to see if the downloaded mp3 already has tags, and if not to set them using the album data.