Support for downloading albums
Although crude.

Next up, pulling the album data from EmbedData instead of TralbumData to remove the need for a check if the URL is to a track or an album, as well as to clean up loading the 'current' as JSON and parsing through that looking for album data.

Side note: it currently doesn't support standalone tracks (as in, tracks that are not part of an album). It still works for standalone tracks that exist within an album.

Also side note: it doesn't work for albums that don't have an explicit artist.